third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-preflight-request-invalid-status-501-expected.txt

CONSOLE ERROR: Failed to load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-request-invalid-status.php?code=501: Response for preflight has invalid HTTP status code 501
Test that preflight requests with invalid status code results in error. Should print PASS.

PASS
